updateIsPublishedForTemplate
