+++ lib/templates/updateIsPublishedForTemplate.ts
@@ -1,0 +1,75 @@
+import { formCache } from "../cache/formCache";
+import { prisma, prismaErrors } from "@gcforms/database";
+import { FormRecord } from "@lib/types";
+import { authorization } from "../privileges";
+import { AuditLogAccessDeniedDetails, logEvent } from "../auditLogs";
+import { deleteDraftFormResponses } from "../vault";
+import { TemplateAlreadyPublishedError } from "./errors";
+import { getFullTemplateByID } from "./getFullTemplateByID";
+import { parseTemplate } from "./shared";
+
+/**
+ * Update `isPublished` value for a specific form.
+ */
+export async function updateIsPublishedForTemplate(
+  formID: string,
+  isPublished: boolean,
+  publishReason: string,
+  publishFormType: string,
+  publishDescription: string
+): Promise<FormRecord | null> {
+  // Alias the isPublished value to newPublishStatus for clarity within the function
+  const newPublishStatus = isPublished;
+
+  const { user } = await authorization.canPublishForm(formID).catch((e) => {
+    logEvent(
+      e.user.id,
+      { type: "Form", id: formID },
+      "AccessDenied",
+      AuditLogAccessDeniedDetails.AccessDenied_AttemptToPublishForm
+    );
+    throw e;
+  });
+
+  // Delete all form responses created during draft mode
+  if (newPublishStatus && process.env.APP_ENV !== "test") {
+    try {
+      await deleteDraftFormResponses(formID);
+    } catch (e) {
+      if (e instanceof TemplateAlreadyPublishedError) {
+        // Already published, so we can just return the full template
+        return getFullTemplateByID(formID, false);
+      }
+
+      throw e;
+    }
+  }
+
+  const updatedTemplate = await prisma.template
+    .update({
+      where: {
+        id: formID,
+        isPublished: {
+          not: newPublishStatus, // Only update if the current publish status is different from the new one,
+        },
+      },
+      data: {
+        isPublished: newPublishStatus,
+        publishReason: publishReason,
+        publishFormType: publishFormType,
+        publishDesc: publishDescription,
+      },
+      include: {
+        deliveryOption: true,
+      },
+    })
+    .catch((e) => prismaErrors(e, null));
+
+  if (updatedTemplate === null) return updatedTemplate;
+
+  if (formCache.cacheAvailable) formCache.invalidate(formID);
+
+  logEvent(user.id, { type: "Form", id: formID }, "PublishForm");
+
+  return parseTemplate(updatedTemplate);
+}
